chitChatBackend/requirements.txt
asgiref
attrs
certifi
Django
django-cors-headers
django-filter
djangorestframework
drf-spectacular
drf-spectacular-sidecar
inflection
jsonschema
Markdown
names
numpy
panda
pyparsing
pyrsistent
python-dateutil
pyt
PyYAML
six
sqlparse
tzdata
uritemplate
wincertstore
pycryptodome
daphne
channels
websocket
websocket_client
djangorestframework_simplejwt
